Add tests from new backend and processor functions
